// src/app/_dto/icvatcat.dto.ts
export class IcvatcatDto {

    coCode: string;

    code: number;

    name: string | null;

    vatPer: number;

    acCode: string | null;

    vatType: string;
  
    sortOrder: number | null;

    status: string;

    sysuser: string;

    sysdate: Date;

    modiuser: string | null;

    modidate: Date | null;
  }